'big' update: I don't save all the data in a Vec<u8> anymore, instead just always piece by piece. Also added a DB in the background which keeps track of the progress

+++ src/messages/payloads.rs
@@ -69,8 +69,8 @@ pub struct ResponsePiecePayload {
 impl Payload for ResponsePiecePayload {
     fn from_be_bytes(bytes: &[u8]) -> Self {
         let block_length = bytes.len() - 8;
-        let mut block = vec![0u8; block_length as usize];
-        block[..block_length as usize].copy_from_slice(&bytes[8..8 + block_length as usize]);
+        let mut block = vec![0u8; block_length];
+        block[..block_length].copy_from_slice(&bytes[8..8 + block_length]);
         Self {
             index: u32::from_be_bytes([bytes[0], bytes[1], bytes[2], bytes[3]]),
             begin: u32::from_be_bytes([bytes[4], bytes[5], bytes[6], bytes[7]]),
@@ -78,7 +78,7 @@ impl Payload for ResponsePiecePayload {
         }
     }
     fn to_be_bytes(&self) -> Vec<u8> {
-        let mut bytes = vec![0u8; 8 + self.block.len() as usize];
+        let mut bytes = vec![0u8; 8 + self.block.len()];
         bytes[0..4].copy_from_slice(&self.index.to_be_bytes());
         bytes[4..8].copy_from_slice(&self.begin.to_be_bytes());
         bytes[8..].copy_from_slice(&self.block);
